requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-bootstrap-modal-forms==2.1.0
django-friendship==1.9.1
django-heroku==0.0.0
django-m2m-history==0.3.6
django-widget-tweaks==1.4.8
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-twitter==3.5
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.0
whitenoise==5.2.0
wrapt==1.12.1
